refactored Home page and made search field responsive to all devices which was not before
